requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
PyJWT==2.6.0
requests==2.28.2
pyaudio==0.2.14
vosk==0.3.45
deepmultilingualpunctuation==1.0.0
pdfplumber==0.9.0
python-pptx==0.6.21
python-docx==0.8.11
librosa==0.10.0
torch>=2.3.1
transformers==4.35.2
google-cloud-firestore==2.7.0
python-multipart==0.0.5
argostranslate==1.9.6
openai==0.28






